cement: fix bug on early notes
If the player hit a note before the song started we would get an error.
Now we are OK with early notes.  We just set t0 before we start the
count (and add the count time to it).  This way if they strike early
they get an accurate score instead of an exception.